+++ src/main/index.js
@@ -134,7 +134,7 @@ app.on('open-url', (event, url) => {
   broadcastURL(deeplinkingUrl)
 })
 
-app.setAsDefaultProtocolClient('ark', process.execPath, ['--'])
+app.setAsDefaultProtocolClient('ock', process.execPath, ['--'])
 
 /**
  * Auto Updater