+++ src/renderer/components/Wallet/WalletDelegates/WalletDelegates.vue
@@ -132,7 +132,7 @@ export default {
     },
 
     votingUrl () {
-      return 'https://docs.ark.io/tutorials/usage-guides/how-to-vote-in-the-ark-desktop-wallet.html'
+      return 'https://docs.ark.io/tutorials/usage-guides/how-to-vote-in-the-ockham-desktop-wallet.html'
     }
   },