+++ src/renderer/i18n/date-time-formats.js
@@ -18,6 +18,14 @@ export default {
       year: 'numeric', month: 'short', day: 'numeric', weekday: 'short', hour: 'numeric', minute: 'numeric'
     }
   },
+  'fr-FR': {
+    short: {
+      year: 'numeric', month: 'short', day: 'numeric'
+    },
+    long: {
+      year: 'numeric', month: 'short', day: 'numeric', weekday: 'short', hour: 'numeric', minute: 'numeric'
+    }
+  },
   'pt-BR': {
     short: {
       year: 'numeric', month: 'short', day: 'numeric'