+++ src/renderer/i18n/locales/pt-BR.js
@@ -3,6 +3,7 @@ export default {
   LANGUAGES: {
     'en-US': 'Inglês',
     'es-ES': 'Espanhol',
+    'fr-FR': 'Francês',
     'pt-BR': 'Português - Brasil'
   }